*breaking* refactor: type mapping changed for `T | Reference Object`
Previously, all types that could be unioned with a Reference Object had a `$ref` field. However those types are also
used standalone in the specification. To bring the type mapping closer to the specification, these types no longuer have
a `$ref` field. Impacted types are `Parameter` and `Response`. To deal with the union case, companion types were added
and named using the pattern `TOrReference`: `ParameterOrReference` and `ResponseOrReference`. These two types extends
their respective standalone type, and add a `$ref` field. It now as easy to distinguish places where references are
accepted and those where they are not as it is in the specification.